Split two files out of the noExcessiveLinesPerFile ignore list
Remove `src/features/public/ticket-submit.ts` and
`src/shared/band-name-generator/word-pools.ts` from the biome
`noExcessiveLinesPerFile` override list, then split each so every resulting
file is well under the 1,000-line ceiling.

- word-pools.ts (1,457 lines of pure data) becomes a `word-pools/` folder of
  five themed files (band-words, band-nouns, music, venues, shows), each under
  400 lines; band-name-generator.ts imports and re-exports from them. Data is
  byte-identical (verified by checksum).
- ticket-submit.ts (1,205 lines) splits into three: `ticket-pricing.ts` (the
  pure pricing/quote helpers), `ticket-process.ts` (the parse-price-complete
  submission pipeline), and `ticket-submit.ts` (the routing/rendering shell).

Splitting ticket-submit.ts let jscpd fully scan it and surfaced a previously
hidden clone: the `ChildCapacityInfo` assembly was duplicated between
renderCtx and discovery.ts. Extract `childCapacityInfo` in discovery.ts and
call it from both sites, restoring 0% duplication.
