+++ biome.json
@@ -154,8 +154,6 @@
       "includes": [
         "src/features/api/index.ts",
         "src/features/api/payment-processing.ts",
-        "src/features/public/ticket-submit.ts",
-        "src/shared/band-name-generator/word-pools.ts",
         "src/shared/db/migrations/schema.ts",
         "src/shared/db/questions.ts",
         "src/shared/db/settings.ts",
