--- src/shared/band-name-generator/word-pools.ts
@@ -1,1457 +1,0 @@
-/**
- * Word pools for the band-name/venue/gig-description generator.
- * Each pool is a flat array of strings; band-name-generator.ts wires
- * them to template slots via SLOT_POOLS.
- */
-
-/** Weird, evocative adjectives suitable for metal/rock band names */
-export const BAND_ADJECTIVES = [
-  "Brutal",
-  "Equine",
-  "Mystic",
-  "Screaming",
-  "Savage",
-  "Eternal",
-  "Forbidden",
-  "Cursed",
-  "Crystal",
-  "Velvet",
-  "Iron",
-  "Black",
-  "Crimson",
-  "Ghastly",
-  "Howling",
-  "Roaring",
-  "Whispering",
-  "Burning",
-  "Frozen",
-  "Molten",
-  "Electric",
-  "Toxic",
-  "Radioactive",
-  "Phantom",
-  "Spectral",
-  "Skeletal",
-  "Demonic",
-  "Angelic",
-  "Holy",
-  "Unholy",
-  "Profane",
-  "Sacred",
-  "Vengeful",
-  "Wretched",
-  "Tragic",
-  "Tortured",
-  "Twisted",
-  "Bleeding",
-  "Pale",
-  "Withered",
-  "Reanimated",
-  "Drunken",
-  "Feral",
-  "Rabid",
-  "Mighty",
-  "Tyrannical",
-  "Imperial",
-  "Royal",
-  "Divine",
-  "Apocalyptic",
-  "Cosmic",
-  "Lunar",
-  "Solar",
-  "Stellar",
-  "Astral",
-  "Voidborn",
-  "Infernal",
-  "Diabolical",
-  "Sinister",
-  "Cataclysmic",
-  "Volcanic",
-  "Glacial",
-  "Granite",
-  "Marbled",
-  "Gilded",
-  "Tarnished",
-  "Mercurial",
-  "Mongolian",
-  "Jurassic",
-  "Chrome",
-  "Diesel",
-  "Atomic",
-  "Neon",
-  "Plastic",
-  "Wooden",
-  "Rusted",
-  "Festering",
-  "Slithering",
-  "Glowing",
-  "Magnetic",
-  "Thunderous",
-  "Subterranean",
-  "Hyperborean",
-  "Perpetual",
-  "Hooded",
-  "Velour",
-  "Magic",
-  "Naked",
-  "Hidden",
-  "Drowning",
-  "Choking",
-  "Greasy",
-  "Dusty",
-  "Rotting",
-  "Yelping",
-  "Polite",
-  "Ancient",
-  "Vintage",
-  "Final",
-  "Last",
-  "Crawling",
-  "Lurking",
-  "Hovering",
-  "Disgraced",
-  "Disowned",
-  "Defeated",
-  "Vanquished",
-  "Resurrected",
-  "Banished",
-  "Exiled",
-  "Wandering",
-  "Lost",
-  "Hunted",
-  "Haunted",
-  "Possessed",
-  "Anointed",
-  "Forsworn",
-  "Forgotten",
-  "Buried",
-  "Entombed",
-  "Enchanted",
-  "Bewitched",
-  "Spellbound",
-  "Hexed",
-  "Doomed",
-  "Damned",
-  "Glorious",
-  "Sublime",
-  "Foul",
-  "Putrid",
-  "Rancid",
-  "Acrid",
-  "Smouldering",
-  "Sizzling",
-  "Roasting",
-  "Boiling",
-  "Stewing",
-  "Bubbling",
-  "Sputtering",
-  "Crackling",
-  "Flickering",
-  "Faltering",
-  "Trembling",
-  "Shuddering",
-  "Quivering",
-  "Quaking",
-  "Wailing",
-  "Snarling",
-  "Gnashing",
-  "Spitting",
-  "Hissing",
-  "Buzzing",
-  "Droning",
-  "Squawking",
-  "Squelching",
-  "Gurgling",
-  "Slurping",
-  "Sticky",
-  "Knotted",
-  "Tangled",
-  "Crooked",
-  "Crumpled",
-  "Crumbling",
-  "Shattered",
-  "Splintered",
-  "Cracked",
-  "Fractured",
-  "Ruined",
-  "Decrepit",
-  "Derelict",
-  "Forlorn",
-  "Bereft",
-  "Abandoned",
-  "Forsaken",
-  "Regrettable",
-  "Ill-Advised",
-  "Untrustworthy",
-  "Suspicious",
-  "Unhinged",
-  "Bewildered",
-  "Befuddled",
-  "Flummoxed",
-  "Ungovernable",
-  "Inconsolable",
-  "Unbothered",
-  "Begrudging",
-  "Intolerable",
-  "Embarrassing",
-  "Implausible",
-  "Improbable",
-  "Unspeakable",
-  "Inexplicable",
-  "Unrepentant",
-  "Belligerent",
-  "Insatiable",
-  "Inevitable",
-  "Indomitable",
-  "Insufferable",
-  "Impervious",
-  "Imperishable",
-  "Incandescent",
-  "Indignant",
-  "Truculent",
-  "Petulant",
-  "Diminutive",
-  "Gargantuan",
-  "Colossal",
-  "Lopsided",
-  "Bedraggled",
-  "Bewhiskered",
-  "Bespectacled",
-  "Cantankerous",
-  "Pernicious",
-  "Lugubrious",
-  "Saccharine",
-  "Melodious",
-  "Pugilistic",
-  "Sclerotic",
-] as const;
-
-/** Audience-eligibility blurbs */
-export const AGE_NOTES = [
-  "all ages welcome",
-  "18+ only",
-  "14+ with parent",
-  "family-friendly",
-  "parental advisory in effect",
-  "strictly grown-ups",
-  "over-30s strongly encouraged",
-  "no under-fives without ear defenders",
-  "pets on leads only",
-  "ID checks at the door",
-  "members and accompanied guests",
-  "carer admitted free",
-  "no babes in arms after 9pm",
-  "well-behaved teens permitted",
-  "responsible adults preferred",
-] as const;
-
-/** Predictable consequences of attending the gig */
-export const AUDIENCE_OUTCOMES = [
-  "regret",
-  "glory",
-  "tinnitus",
-  "sweat",
-  "tears",
-  "mild concern",
-  "confusion",
-  "broken eardrums",
-  "lost dignity",
-  "hoarse voices",
-  "lost car keys",
-  "aching necks",
-  "mismatched footwear",
-  "sticky floors",
-  "questionable bruises",
-  "slightly damp denim",
-  "sudden tinnitus",
-  "minor brain fog",
-  "unexplained leaves in pockets",
-  "inappropriate enthusiasm",
-  "newfound dietary requirements",
-  "missed last buses",
-  "incoherent stories",
-  "second-hand glitter",
-] as const;
-
-/** Collective nouns for "the band's reputation" */
-export const BAND_DESCRIPTORS = [
-  "pioneers",
-  "warriors",
-  "legends",
-  "veterans",
-  "prophets",
-  "godfathers",
-  "stalwarts",
-  "diehards",
-  "originators",
-  "cult heroes",
-  "mavericks",
-  "renegades",
-  "outlaws",
-  "old hands",
-  "elder statesmen",
-  "scene leaders",
-  "underdogs",
-  "scene-stealers",
-  "deep-cut specialists",
-  "self-styled visionaries",
-  "improbable survivors",
-  "eternal underdogs",
-  "long-suffering road dogs",
-  "near-mythical figures",
-  "barely-functional collective",
-  "loosely-affiliated crew",
-  "occasionally-touring side project",
-  "perennially-overlooked outfit",
-  "criminally-undercelebrated trio",
-  "begrudgingly-respected veterans",
-] as const;
-
-/** Verbs the band itself does — used in "{bandDescriptor} {bandVerb} ..." patterns */
-export const BAND_VERBS = [
-  "reunite",
-  "return",
-  "regroup",
-  "reform",
-  "ascend",
-  "descend",
-  "emerge",
-  "materialise",
-  "awaken",
-  "surface",
-  "converge",
-  "rally",
-  "assemble",
-  "manifest",
-  "reappear",
-  "convene",
-  "reconvene",
-  "stagger back",
-  "shamble forth",
-  "limp onstage",
-  "march in",
-  "stomp out",
-  "drift in",
-  "crash through the door",
-] as const;
-
-/** Adjective phrases describing the state of a "weird venue" before being repurposed */
-export const BUILDING_STATES = [
-  "converted",
-  "repurposed",
-  "abandoned",
-  "decommissioned",
-  "gentrified",
-  "restored",
-  "haunted",
-  "condemned",
-  "derelict",
-  "half-finished",
-  "recently-evacuated",
-  "formerly-respectable",
-  "barely-legal",
-  "improbably-listed",
-  "informally-occupied",
-  "structurally-unsound",
-  "reluctantly-available",
-  "council-owned",
-  "privately-disowned",
-  "long-disused",
-  "freshly-tagged",
-  "improvisationally-rewired",
-  "inadvisably-reopened",
-] as const;
-
-/** Prepositional phrases connecting clauses without being entirely hard-coded */
-export const CONNECTORS = [
-  "with",
-  "featuring",
-  "accompanied by",
-  "fuelled by",
-  "propelled by",
-  "soundtracked by",
-  "garnished with",
-  "interrupted by",
-  "haunted by",
-  "supported by",
-  "underwritten by",
-  "invigilated by",
-  "officiated by",
-  "narrated by",
-  "punctuated by",
-  "occasionally upstaged by",
-  "loosely supervised by",
-  "barely tolerated by",
-] as const;
-
-/** Themes for crossover/novelty acts */
-export const CROSSOVERS = [
-  "sea-shanty",
-  "druid",
-  "viking",
-  "pirate",
-  "robot",
-  "ghost",
-  "morris-dance",
-  "bagpipe",
-  "yodelling",
-  "yacht-rock",
-  "monster-truck",
-  "sci-fi",
-  "wild-west",
-  "Tudor-court",
-  "Edwardian-music-hall",
-  "Roman-legion",
-  "samurai",
-  "lumberjack",
-  "tea-party",
-  "coal-miner",
-  "deep-sea-trawler",
-  "haunted-circus",
-  "post-apocalyptic",
-  "Cold-War",
-  "Steampunk",
-  "Cyberpunk",
-  "renaissance-fair",
-  "1970s-prog",
-  "1980s-disco",
-  "1990s-rave",
-  "early-noughties-emo",
-  "small-claims-court",
-  "town-council",
-  "garden-centre",
-] as const;
-
-/** What the listing itself is — sits where words like "show" / "matinee" used to be hard-coded */
-export const LISTING_TYPES = [
-  "night",
-  "matinee",
-  "show",
-  "evening",
-  "gig",
-  "set",
-  "recital",
-  "séance",
-  "summoning",
-  "ritual",
-  "vigil",
-  "broadcast",
-  "takeover",
-  "residency",
-  "occupation",
-  "livestream",
-  "riot",
-  "jamboree",
-  "knees-up",
-  "happening",
-  "spectacle",
-  "exhibition",
-  "convocation",
-  "wake",
-  "send-off",
-  "reckoning",
-  "exorcism",
-] as const;
-
-/** Outdoor / listing-style venue archetypes */
-export const FESTIVAL_TYPES = [
-  "Festival",
-  "Fest",
-  "Fair",
-  "Carnival",
-  "Convention",
-  "Summit",
-  "Showcase",
-  "Massacre",
-  "Slaughter",
-  "Bash",
-  "Jamboree",
-  "Riot",
-  "Reckoning",
-  "Pilgrimage",
-  "Procession",
-  "Reunion",
-  "Tour",
-  "Crusade",
-  "Onslaught",
-  "Weekender",
-  "All-Dayer",
-  "Knees-Up",
-  "Hoedown",
-  "Stomp",
-  "Throwdown",
-  "Free-For-All",
-  "Marathon",
-  "Wakefield",
-  "Vigil",
-  "Spectacle",
-  "Extravaganza",
-  "Gala",
-  "Soirée",
-  "Roadshow",
-  "Symposium",
-  "Conclave",
-  "Assembly",
-  "Gathering",
-  "Reunion Tour",
-  "Exhibition",
-  "Tournament",
-  "Olympiad",
-] as const;
-
-/** Sub-genres used inside gig-description templates */
-export const GENRES = [
-  "doom",
-  "sludge",
-  "prog",
-  "hardcore",
-  "glam",
-  "thrash",
-  "death-metal",
-  "black-metal",
-  "stoner-rock",
-  "folk-metal",
-  "industrial",
-  "symphonic",
-  "hair-metal",
-  "gothic",
-  "drone",
-  "speed-metal",
-  "power-metal",
-  "psychobilly",
-  "darkwave",
-  "post-punk",
-  "shoegaze",
-  "grindcore",
-  "blackened-jazz",
-  "viking-metal",
-  "pirate-metal",
-  "doom-jazz",
-  "djent",
-  "crust-punk",
-  "dream-pop",
-  "math-rock",
-  "noise-rock",
-  "krautrock",
-  "post-rock",
-  "post-metal",
-  "blackgaze",
-  "deathcore",
-  "metalcore",
-  "groovecore",
-  "nu-metal",
-  "rap-metal",
-  "country-metal",
-  "celtic-metal",
-  "medieval-metal",
-  "baroque-metal",
-  "circus-metal",
-  "chiptune-metal",
-  "synthwave",
-  "electroclash",
-  "horror-punk",
-  "cowpunk",
-  "skiffle-metal",
-  "dungeon-synth",
-  "bedroom-doom",
-  "garage-thrash",
-  "polka-grind",
-  "screamo",
-  "emo-violence",
-  "math-core",
-  "post-hardcore",
-  "djent-folk",
-  "chamber-doom",
-  "zydeco-thrash",
-  "sea-shanty-core",
-] as const;
-
-/** Hyperbolic adjective phrases for sensory descriptions */
-export const INTENSITIES = [
-  "ear-splitting",
-  "brain-melting",
-  "head-splitting",
-  "ground-shaking",
-  "soul-crushing",
-  "deafening",
-  "earth-shattering",
-  "skin-peeling",
-  "bone-rattling",
-  "chest-thumping",
-  "thoroughly inadvisable",
-  "barely-legal",
-  "trouser-flapping",
-  "hair-raising",
-  "spleen-loosening",
-  "neighbour-bothering",
-  "windscreen-shattering",
-  "pavement-cracking",
-  "ceiling-loosening",
-  "double-glazing-defeating",
-  "rib-rattling",
-  "cardigan-blowing",
-  "biscuit-tin-disturbing",
-  "regrettably enthusiastic",
-  "unreasonably cheerful",
-  "borderline-dangerous",
-  "questionably-permitted",
-  "absurdly-loud",
-  "unaccountably-aggressive",
-  "knee-buckling",
-  "teeth-shaking",
-  "moustache-curling",
-  "ozone-depleting",
-  "hedge-trimming",
-  "impressively-illegal",
-] as const;
-
-/** Verbs for "the band ___ed the venue" */
-export const NOISE_VERBS = [
-  "invade",
-  "descend on",
-  "tear through",
-  "assault",
-  "demolish",
-  "conquer",
-  "devastate",
-  "obliterate",
-  "occupy",
-  "set fire to",
-  "lay siege to",
-  "blockade",
-  "annex",
-  "decommission",
-  "rearrange",
-  "renegotiate",
-  "remodel",
-  "reupholster",
-  "redecorate",
-  "fumigate",
-  "rezone",
-  "dismantle",
-  "reassemble",
-  "abscond from",
-  "infiltrate",
-  "depopulate",
-  "haunt",
-  "stalk",
-  "loiter in",
-  "graffiti",
-  "monopolise",
-  "outlaw",
-] as const;
-
-/** Concrete and abstract nouns that work as band-name subjects */
-export const BAND_NOUNS = [
-  "Fox",
-  "Garlic",
-  "Handshake",
-  "Hand",
-  "Wolf",
-  "Raven",
-  "Crow",
-  "Vulture",
-  "Spider",
-  "Scorpion",
-  "Serpent",
-  "Dragon",
-  "Hydra",
-  "Kraken",
-  "Octopus",
-  "Shark",
-  "Whale",
-  "Phoenix",
-  "Griffin",
-  "Behemoth",
-  "Tomb",
-  "Tower",
-  "Crypt",
-  "Cathedral",
-  "Temple",
-  "Throne",
-  "Crown",
-  "Skull",
-  "Bone",
-  "Blood",
-  "Fang",
-  "Claw",
-  "Wing",
-  "Eye",
-  "Heart",
-  "Spine",
-  "Mind",
-  "Soul",
-  "Ghost",
-  "Witch",
-  "Wizard",
-  "Warlock",
-  "Necromancer",
-  "Knight",
-  "Crusader",
-  "Marauder",
-  "Berserker",
-  "Viking",
-  "Templar",
-  "Reaper",
-  "Executioner",
-  "Hangman",
-  "Butcher",
-  "Doctor",
-  "Surgeon",
-  "Priest",
-  "Prophet",
-  "Oracle",
-  "Pilgrim",
-  "Wanderer",
-  "Vagabond",
-  "Mariner",
-  "Pirate",
-  "Monarch",
-  "Tyrant",
-  "Emperor",
-  "Beast",
-  "Hound",
-  "Stallion",
-  "Mongoose",
-  "Otter",
-  "Falcon",
-  "Eagle",
-  "Owl",
-  "Bat",
-  "Moth",
-  "Beetle",
-  "Toad",
-  "Frog",
-  "Newt",
-  "Cobra",
-  "Mantis",
-  "Hornet",
-  "Wasp",
-  "Locust",
-  "Cicada",
-  "Salmon",
-  "Pike",
-  "Eel",
-  "Squid",
-  "Tree",
-  "Forest",
-  "Mountain",
-  "River",
-  "Ocean",
-  "Storm",
-  "Tempest",
-  "Cyclone",
-  "Avalanche",
-  "Comet",
-  "Meteor",
-  "Nebula",
-  "Galaxy",
-  "Void",
-  "Abyss",
-  "Inferno",
-  "Pit",
-  "Cavern",
-  "Dungeon",
-  "Castle",
-  "Fortress",
-  "Citadel",
-  "Bastion",
-  "Hammer",
-  "Axe",
-  "Sword",
-  "Dagger",
-  "Mace",
-  "Whip",
-  "Chain",
-  "Anvil",
-  "Forge",
-  "Cauldron",
-  "Chalice",
-  "Goblet",
-  "Mirror",
-  "Compass",
-  "Lantern",
-  "Candle",
-  "Coffin",
-  "Casket",
-  "Sarcophagus",
-  "Pyre",
-  "Procession",
-  "Gallery",
-  "Pantheon",
-  "Rabbit",
-  "Badger",
-  "Goose",
-  "Duck",
-  "Pigeon",
-  "Sandwich",
-  "Onion",
-  "Turnip",
-  "Cabbage",
-  "Pickle",
-  "Trousers",
-  "Moustache",
-  "Lighthouse",
-  "Bunker",
-  "Library",
-  "Aqueduct",
-  "Catacomb",
-  "Mausoleum",
-  "Reliquary",
-  "Reservoir",
-  "Telegram",
-  "Manuscript",
-  "Treatise",
-  "Bestiary",
-  "Atlas",
-  "Almanac",
-  "Codex",
-  "Compendium",
-  "Calliope",
-  "Carousel",
-  "Caboose",
-  "Pram",
-  "Ladder",
-  "Trapdoor",
-  "Greenhouse",
-  "Chandelier",
-  "Carpet",
-  "Wallpaper",
-  "Doorknob",
-  "Letterbox",
-  "Bicycle",
-  "Tractor",
-  "Combine",
-  "Submarine",
-  "Zeppelin",
-  "Helicopter",
-  "Steamboat",
-  "Bobsleigh",
-  "Pomegranate",
-  "Tangerine",
-  "Persimmon",
-  "Aubergine",
-  "Courgette",
-  "Pineapple",
-  "Apricot",
-  "Truffle",
-  "Goulash",
-  "Custard",
-  "Marmalade",
-  "Brie",
-  "Pavlova",
-  "Caviar",
-  "Schnitzel",
-  "Vol-au-Vent",
-  "Mousse",
-  "Spaniel",
-  "Greyhound",
-  "Llama",
-  "Capybara",
-  "Walrus",
-  "Manatee",
-  "Pangolin",
-  "Aardvark",
-  "Wombat",
-  "Gibbon",
-  "Toucan",
-  "Pelican",
-  "Cormorant",
-  "Heron",
-  "Stork",
-  "Magpie",
-  "Albatross",
-  "Penguin",
-  "Anglerfish",
-  "Stingray",
-  "Lobster",
-  "Cuttlefish",
-  "Nautilus",
-  "Periwinkle",
-  "Beanstalk",
-  "Mushroom",
-  "Toadstool",
-  "Bramble",
-  "Thistle",
-  "Bracken",
-  "Yew",
-  "Hawthorn",
-  "Elderflower",
-  "Comfrey",
-  "Mandrake",
-  "Wormwood",
-  "Bellringer",
-  "Innkeeper",
-  "Postmaster",
-  "Tinker",
-  "Cobbler",
-  "Coachman",
-  "Stableboy",
-  "Magistrate",
-  "Footpad",
-  "Highwayman",
-] as const;
-
-/** Spelled-out small numbers for descriptions */
-export const NUMBER_WORDS = [
-  "two",
-  "three",
-  "four",
-  "five",
-  "six",
-  "seven",
-  "eight",
-  "nine",
-  "ten",
-  "eleven",
-  "twelve",
-  "thirteen",
-  "slistingeen",
-  "twenty-one",
-  "thirty",
-  "an alarming number of",
-  "an undisclosed number of",
-  "no fewer than four",
-  "approximately seven",
-  "exactly one",
-] as const;
-
-/** Wildly out-of-place instruments for novelty descriptions */
-export const ODD_INSTRUMENTS = [
-  "theremin",
-  "ukulele",
-  "bagpipes",
-  "hurdy-gurdy",
-  "kazoo",
-  "glockenspiel",
-  "accordion",
-  "harp",
-  "sitar",
-  "mandolin",
-  "stylophone",
-  "didgeridoo",
-  "vacuum cleaner",
-  "Casio keyboard",
-  "harpsichord",
-  "calliope",
-  "musical saw",
-  "spoons",
-  "washboard",
-  "jaw harp",
-  "ocarina",
-  "pan flute",
-  "alphorn",
-  "glass harmonica",
-  "lap steel",
-  "slide whistle",
-  "swanee whistle",
-  "wobble board",
-  "Stylophone Beatbox",
-  "single bagpipe drone",
-  "broken Wurlitzer",
-  "ironing board with strings",
-  "rubber-band orchestra",
-  "tuned biscuit tin",
-  "borrowed cello",
-  "untuned banjo",
-  "kettle and spoon",
-  "vintage ondes Martenot",
-  "bicycle bell choir",
-  "amplified beard",
-] as const;
-
-/** First names used for "{Adjective} {Name} and Friends"-style patterns */
-export const BAND_PERSON_NAMES = [
-  "David",
-  "Ozzy",
-  "Lemmy",
-  "Dio",
-  "Axl",
-  "Slash",
-  "Iggy",
-  "Bowie",
-  "Bruce",
-  "Eddie",
-  "Glenn",
-  "Henry",
-  "Tommy",
-  "Johnny",
-  "Ronnie",
-  "Vincent",
-  "Alice",
-  "Stevie",
-  "Jimmy",
-  "Roger",
-  "Trevor",
-  "Brian",
-  "Geoff",
-  "Maureen",
-  "Pamela",
-  "Cassandra",
-  "Susan",
-  "Beverly",
-  "Gladys",
-  "Doris",
-  "Mildred",
-  "Wilma",
-  "Bernard",
-  "Norman",
-  "Reginald",
-  "Cyril",
-  "Mortimer",
-  "Algernon",
-  "Cornelius",
-  "Bartholomew",
-  "Archibald",
-  "Percival",
-  "Horatio",
-  "Egbert",
-  "Lionel",
-  "Clive",
-  "Roland",
-  "Nigel",
-  "Wilbur",
-  "Cuthbert",
-  "Humphrey",
-  "Cedric",
-  "Stanley",
-  "Walter",
-  "Sidney",
-  "Herbert",
-  "Edmund",
-  "Reuben",
-  "Otto",
-  "Fritz",
-  "Boris",
-  "Igor",
-  "Wolfgang",
-  "Heinrich",
-  "Ingrid",
-  "Astrid",
-  "Brunhilda",
-  "Helga",
-  "Ursula",
-  "Greta",
-  "Magda",
-  "Edith",
-  "Ethel",
-  "Hilda",
-  "Beatrice",
-  "Constance",
-  "Prudence",
-  "Patience",
-  "Tabitha",
-  "Agatha",
-  "Hortense",
-  "Drusilla",
-  "Lavinia",
-  "Octavia",
-  "Esmeralda",
-  "Mavis",
-  "Dolores",
-  "Phyllis",
-  "Vera",
-  "Mabel",
-  "Pearl",
-  "Marjorie",
-  "Gertrude",
-  "Lucille",
-  "Verity",
-  "Cordelia",
-  "Imelda",
-  "Reinhold",
-  "Bjorn",
-  "Sven",
-  "Ulrich",
-  "Klaus",
-  "Hans",
-  "Dieter",
-] as const;
-
-/** Karaoke / dress-up prize categories */
-export const PRIZES = [
-  "biggest hair",
-  "loudest scream",
-  "best devil horns",
-  "most convincing growl",
-  "longest air-guitar solo",
-  "deepest voice",
-  "most denim worn at once",
-  "most authentic mullet",
-  "most enthusiastic windmill arm",
-  "best fake corpse-paint",
-  "loudest unaccompanied howl",
-  "longest sustained note",
-  "most accurate Lemmy impression",
-  "fastest tambourine work",
-  "most dramatic hair flip",
-  "most-improved scream since last time",
-  "best-disguised vocal warm-up",
-  "most committed knee-slide",
-  "most emphatic horns-up gesture",
-  "best Joe-Cocker-style face",
-] as const;
-
-/** Things that are absolutely not allowed at this gig */
-export const PROHIBITIONS = [
-  "stage diving onto the pizza",
-  "phones above shoulder height",
-  "vuvuzelas",
-  "glitter cannons",
-  "throwing pints",
-  "uninvited cowbell",
-  "amateur pyrotechnics",
-  "crowdsurfing inflatables",
-  "audience requests",
-  "pretending to know the lyrics",
-  "shouting Wonderwall",
-  "encore-baiting",
-  "bringing your own setlist",
-  "open-toed footwear in the pit",
-  "drum solos by attendees",
-  "tribute-act T-shirts",
-  "loud crisp packets",
-  "moshing during ballads",
-  "talking during ballads",
-  "filming with two phones at once",
-  "selfie sticks of any length",
-  "questioning the headliner's setlist",
-  "improvised lyrics",
-  "queue-jumping at merch",
-  "uninvited stage invasions",
-  "balloon animals",
-  "vapes the size of a kettle",
-  "outdoor flares",
-  "professional cameras without a pass",
-  "harmonicas brought from home",
-  "negotiating refunds mid-set",
-  "audible sighs during solos",
-  "inappropriately enthusiastic kazoo work",
-] as const;
-
-/** Items of clothing or attire that are notionally compulsory */
-export const REQUIREMENTS = [
-  "sequins",
-  "leather",
-  "denim",
-  "mascara",
-  "hairspray",
-  "shoulder pads",
-  "bullet belts",
-  "tasselled jackets",
-  "shredded jeans",
-  "spandex",
-  "corpse paint",
-  "studded gauntlets",
-  "pirate hats",
-  "viking helmets",
-  "novelty cravats",
-  "ill-fitting capes",
-  "platform boots",
-  "bandanas (any colour)",
-  "ironic moustaches",
-  "elaborate eyeliner",
-  "battle vests",
-  "back patches",
-  "fingerless gloves",
-  "wraparound sunglasses",
-  "kohl-rimmed eyes",
-  "dramatic neckwear",
-  "regrettable haircuts",
-  "tour T-shirts (preferably someone else's)",
-] as const;
-
-/** Concrete things you'd see / smell / regret at a gig */
-export const SHOW_ITEMS = [
-  "denim",
-  "leather",
-  "mead",
-  "regret",
-  "glory",
-  "smoke",
-  "lasers",
-  "dry ice",
-  "bandanas",
-  "studs",
-  "shredding solos",
-  "blast beats",
-  "double kicks",
-  "fog machines",
-  "questionable lyrics",
-  "obscure cover versions",
-  "extended drum solos",
-  "merch you'll regret on Monday",
-  "improvised pyrotechnics",
-  "spilled beer",
-  "spandex",
-  "corpse paint",
-  "bullet belts",
-  "leather kilts",
-  "body glitter",
-  "patchouli",
-  "warm lager",
-  "cold sausage rolls",
-  "merchandise queues",
-  "guitar picks",
-  "drumstick souvenirs",
-  "a single rogue bagpipe",
-  "exhausted sound engineers",
-  "questionable hairlines",
-  "ill-advised tattoos",
-  "second-hand smoke",
-  "third-hand stories",
-  "sweat-stained setlists",
-  "audible feedback",
-  "inaudible vocals",
-  "broken cymbals",
-  "duct tape",
-  "split jeans",
-  "torn t-shirts",
-  "missing teeth",
-  "discounted earplugs",
-  "questionable life choices",
-  "everyone's last shred of dignity",
-  "free posters",
-  "expired pizza vouchers",
-  "improvised mosh choreography",
-] as const;
-
-/** Tail phrases that turn a noun phrase into a fully-realised band name */
-export const BAND_SUFFIXES = [
-  "of Doom",
-  "of Mercy",
-  "of Sorrow",
-  "of the Damned",
-  "of the Forsaken",
-  "and Friends",
-  "Reunited",
-  "Unleashed",
-  "Resurrected",
-  "Returns",
-  "Forever",
-  "Reborn",
-  "Triumphant",
-  "Ascendant",
-  "of the Apocalypse",
-  "of the Void",
-  "of the Storm",
-  "of the Abyss",
-  "of the North",
-  "of Chaos",
-  "of Madness",
-  "of Eternity",
-  "of the Crypt",
-  "and the Disciples",
-  "and the Acolytes",
-  "and the Pilgrims",
-  "and Sons",
-  "and Daughters",
-  "Experience",
-  "Project",
-  "Collective",
-  "Brigade",
-  "Brotherhood",
-  "Sisterhood",
-  "Cult",
-  "Coven",
-  "Convocation",
-  "Conspiracy",
-  "Manifesto",
-  "Revival",
-  "Live in Concert",
-  "MMXXVI",
-  "of the South",
-  "of the East",
-  "of the West",
-  "of Yesterday",
-  "of Tomorrow",
-  "of the Underworld",
-  "of the Overworld",
-  "of the Marshlands",
-  "of the Highlands",
-  "of the Lowlands",
-  "of the Swamp",
-  "of the Steppe",
-  "of the Tundra",
-  "of the Ravine",
-  "Reanimated",
-  "Refurbished",
-  "Recommissioned",
-  "Reconvened",
-  "Re-Released",
-  "and the Quartermasters",
-  "and the Apothecaries",
-  "and the Notaries",
-  "and the Clerks",
-  "and the Custodians",
-  "Quintet",
-  "Sextet",
-  "Septet",
-  "Orchestra",
-  "Ensemble",
-  "Society",
-  "Guild",
-  "Syndicate",
-  "Authority",
-  "Committee",
-  "Tribunal",
-  "Inquest",
-  "Convention",
-  "Roadshow",
-  "Caravan",
-  "Pageant",
-  "Processional",
-  "Wake",
-  "Vigil",
-] as const;
-
-/** Common gig start times (24h-style hours) */
-export const TIMES = [
-  "6",
-  "7",
-  "7:30",
-  "8",
-  "8:30",
-  "9",
-  "9:30",
-  "10",
-  "11",
-  "midnight",
-] as const;
-
-/** Adjectives that qualify a "tour" or one-off comeback show */
-export const TOUR_ADJECTIVES = [
-  "farewell",
-  "reunion",
-  "comeback",
-  "retrospective",
-  "anniversary",
-  "definitely-the-last",
-  "victory",
-  "redemption",
-  "vengeance",
-  "valedictory",
-  "absolutely-final",
-  "really-really-last",
-  "we-mean-it-this-time",
-  "first and last",
-  "swansong",
-  "homecoming",
-  "tearful",
-  "long-awaited",
-  "long-promised",
-  "long-overdue",
-  "mostly-original-lineup",
-  "approximately-original-lineup",
-  "one-night-only",
-  "two-nights-only",
-  "open-ended",
-] as const;
-
-/** Indoor venue archetypes — concrete buildings/rooms */
-export const VENUE_TYPES = [
-  "Arena",
-  "Academy",
-  "Stadium",
-  "Auditorium",
-  "Hall",
-  "Theatre",
-  "Coliseum",
-  "Amphitheatre",
-  "Pavilion",
-  "Forum",
-  "Dome",
-  "Centre",
-  "Roundhouse",
-  "Tabernacle",
-  "Sanctuary",
-  "Crypt",
-  "Vault",
-  "Bunker",
-  "Warehouse",
-  "Factory",
-  "Mill",
-  "Foundry",
-  "Quarry",
-  "Depot",
-  "Garage",
-  "Lounge",
-  "Club",
-  "Den",
-  "Pit",
-  "Chamber",
-  "Ballroom",
-  "Workingmen's Club",
-  "Social Club",
-  "Working Men's Institute",
-  "Bingo Hall",
-  "Snooker Hall",
-  "Boxing Gym",
-  "Scout Hut",
-  "Village Hall",
-  "Lecture Theatre",
-  "Cinema",
-  "Skating Rink",
-  "Roller Disco",
-  "Bowling Alley",
-  "Mausoleum",
-  "Crematorium",
-  "Conservatory",
-  "Pump House",
-  "Boiler Room",
-  "Engine Shed",
-  "Tram Shed",
-  "Toll House",
-  "Custom House",
-  "Drill Hall",
-  "Masonic Lodge",
-  "Reading Room",
-  "Tea Rooms",
-  "Assembly Rooms",
-  "Music Room",
-  "Practice Room",
-  "Rehearsal Studio",
-] as const;
-
-/** Improbable buildings to be "converted into" a venue */
-export const WEIRD_VENUES = [
-  "lift shaft",
-  "granary",
-  "bus depot",
-  "brewery vat",
-  "postal sorting office",
-  "cattle market",
-  "abandoned multi-storey",
-  "decommissioned chapel",
-  "1980s leisure centre",
-  "shipping container",
-  "village butcher's freezer",
-  "art-deco cinema",
-  "Victorian bathhouse",
-  "Edwardian bandstand",
-  "disused fire station",
-  "former post office",
-  "redundant church",
-  "out-of-service ferry",
-  "decommissioned lighthouse",
-  "abandoned underground station",
-  "WW2 pillbox",
-  "converted barn",
-  "haunted greenhouse",
-  "old reading room",
-  "empty Woolworths",
-  "former Blockbuster",
-  "former HMV",
-  "shut-down bowling alley",
-  "padlocked snooker hall",
-  "boarded-up pub",
-  "derelict library",
-  "scrapped trawler",
-  "retired narrowboat",
-  "carriage-works",
-  "pump house",
-  "abandoned railway tunnel",
-  "long-suffering scout hut",
-  "regrettable nightclub",
-  "decommissioned petrol station",
-] as const;
