+++ src/shared/band-name-generator/word-pools/band-nouns.ts
@@ -1,0 +1,360 @@
+/** Nouns and person names that anchor generated band names. */
+
+/** Concrete and abstract nouns that work as band-name subjects */
+export const BAND_NOUNS = [
+  "Fox",
+  "Garlic",
+  "Handshake",
+  "Hand",
+  "Wolf",
+  "Raven",
+  "Crow",
+  "Vulture",
+  "Spider",
+  "Scorpion",
+  "Serpent",
+  "Dragon",
+  "Hydra",
+  "Kraken",
+  "Octopus",
+  "Shark",
+  "Whale",
+  "Phoenix",
+  "Griffin",
+  "Behemoth",
+  "Tomb",
+  "Tower",
+  "Crypt",
+  "Cathedral",
+  "Temple",
+  "Throne",
+  "Crown",
+  "Skull",
+  "Bone",
+  "Blood",
+  "Fang",
+  "Claw",
+  "Wing",
+  "Eye",
+  "Heart",
+  "Spine",
+  "Mind",
+  "Soul",
+  "Ghost",
+  "Witch",
+  "Wizard",
+  "Warlock",
+  "Necromancer",
+  "Knight",
+  "Crusader",
+  "Marauder",
+  "Berserker",
+  "Viking",
+  "Templar",
+  "Reaper",
+  "Executioner",
+  "Hangman",
+  "Butcher",
+  "Doctor",
+  "Surgeon",
+  "Priest",
+  "Prophet",
+  "Oracle",
+  "Pilgrim",
+  "Wanderer",
+  "Vagabond",
+  "Mariner",
+  "Pirate",
+  "Monarch",
+  "Tyrant",
+  "Emperor",
+  "Beast",
+  "Hound",
+  "Stallion",
+  "Mongoose",
+  "Otter",
+  "Falcon",
+  "Eagle",
+  "Owl",
+  "Bat",
+  "Moth",
+  "Beetle",
+  "Toad",
+  "Frog",
+  "Newt",
+  "Cobra",
+  "Mantis",
+  "Hornet",
+  "Wasp",
+  "Locust",
+  "Cicada",
+  "Salmon",
+  "Pike",
+  "Eel",
+  "Squid",
+  "Tree",
+  "Forest",
+  "Mountain",
+  "River",
+  "Ocean",
+  "Storm",
+  "Tempest",
+  "Cyclone",
+  "Avalanche",
+  "Comet",
+  "Meteor",
+  "Nebula",
+  "Galaxy",
+  "Void",
+  "Abyss",
+  "Inferno",
+  "Pit",
+  "Cavern",
+  "Dungeon",
+  "Castle",
+  "Fortress",
+  "Citadel",
+  "Bastion",
+  "Hammer",
+  "Axe",
+  "Sword",
+  "Dagger",
+  "Mace",
+  "Whip",
+  "Chain",
+  "Anvil",
+  "Forge",
+  "Cauldron",
+  "Chalice",
+  "Goblet",
+  "Mirror",
+  "Compass",
+  "Lantern",
+  "Candle",
+  "Coffin",
+  "Casket",
+  "Sarcophagus",
+  "Pyre",
+  "Procession",
+  "Gallery",
+  "Pantheon",
+  "Rabbit",
+  "Badger",
+  "Goose",
+  "Duck",
+  "Pigeon",
+  "Sandwich",
+  "Onion",
+  "Turnip",
+  "Cabbage",
+  "Pickle",
+  "Trousers",
+  "Moustache",
+  "Lighthouse",
+  "Bunker",
+  "Library",
+  "Aqueduct",
+  "Catacomb",
+  "Mausoleum",
+  "Reliquary",
+  "Reservoir",
+  "Telegram",
+  "Manuscript",
+  "Treatise",
+  "Bestiary",
+  "Atlas",
+  "Almanac",
+  "Codex",
+  "Compendium",
+  "Calliope",
+  "Carousel",
+  "Caboose",
+  "Pram",
+  "Ladder",
+  "Trapdoor",
+  "Greenhouse",
+  "Chandelier",
+  "Carpet",
+  "Wallpaper",
+  "Doorknob",
+  "Letterbox",
+  "Bicycle",
+  "Tractor",
+  "Combine",
+  "Submarine",
+  "Zeppelin",
+  "Helicopter",
+  "Steamboat",
+  "Bobsleigh",
+  "Pomegranate",
+  "Tangerine",
+  "Persimmon",
+  "Aubergine",
+  "Courgette",
+  "Pineapple",
+  "Apricot",
+  "Truffle",
+  "Goulash",
+  "Custard",
+  "Marmalade",
+  "Brie",
+  "Pavlova",
+  "Caviar",
+  "Schnitzel",
+  "Vol-au-Vent",
+  "Mousse",
+  "Spaniel",
+  "Greyhound",
+  "Llama",
+  "Capybara",
+  "Walrus",
+  "Manatee",
+  "Pangolin",
+  "Aardvark",
+  "Wombat",
+  "Gibbon",
+  "Toucan",
+  "Pelican",
+  "Cormorant",
+  "Heron",
+  "Stork",
+  "Magpie",
+  "Albatross",
+  "Penguin",
+  "Anglerfish",
+  "Stingray",
+  "Lobster",
+  "Cuttlefish",
+  "Nautilus",
+  "Periwinkle",
+  "Beanstalk",
+  "Mushroom",
+  "Toadstool",
+  "Bramble",
+  "Thistle",
+  "Bracken",
+  "Yew",
+  "Hawthorn",
+  "Elderflower",
+  "Comfrey",
+  "Mandrake",
+  "Wormwood",
+  "Bellringer",
+  "Innkeeper",
+  "Postmaster",
+  "Tinker",
+  "Cobbler",
+  "Coachman",
+  "Stableboy",
+  "Magistrate",
+  "Footpad",
+  "Highwayman",
+] as const;
+
+/** First names used for "{Adjective} {Name} and Friends"-style patterns */
+export const BAND_PERSON_NAMES = [
+  "David",
+  "Ozzy",
+  "Lemmy",
+  "Dio",
+  "Axl",
+  "Slash",
+  "Iggy",
+  "Bowie",
+  "Bruce",
+  "Eddie",
+  "Glenn",
+  "Henry",
+  "Tommy",
+  "Johnny",
+  "Ronnie",
+  "Vincent",
+  "Alice",
+  "Stevie",
+  "Jimmy",
+  "Roger",
+  "Trevor",
+  "Brian",
+  "Geoff",
+  "Maureen",
+  "Pamela",
+  "Cassandra",
+  "Susan",
+  "Beverly",
+  "Gladys",
+  "Doris",
+  "Mildred",
+  "Wilma",
+  "Bernard",
+  "Norman",
+  "Reginald",
+  "Cyril",
+  "Mortimer",
+  "Algernon",
+  "Cornelius",
+  "Bartholomew",
+  "Archibald",
+  "Percival",
+  "Horatio",
+  "Egbert",
+  "Lionel",
+  "Clive",
+  "Roland",
+  "Nigel",
+  "Wilbur",
+  "Cuthbert",
+  "Humphrey",
+  "Cedric",
+  "Stanley",
+  "Walter",
+  "Sidney",
+  "Herbert",
+  "Edmund",
+  "Reuben",
+  "Otto",
+  "Fritz",
+  "Boris",
+  "Igor",
+  "Wolfgang",
+  "Heinrich",
+  "Ingrid",
+  "Astrid",
+  "Brunhilda",
+  "Helga",
+  "Ursula",
+  "Greta",
+  "Magda",
+  "Edith",
+  "Ethel",
+  "Hilda",
+  "Beatrice",
+  "Constance",
+  "Prudence",
+  "Patience",
+  "Tabitha",
+  "Agatha",
+  "Hortense",
+  "Drusilla",
+  "Lavinia",
+  "Octavia",
+  "Esmeralda",
+  "Mavis",
+  "Dolores",
+  "Phyllis",
+  "Vera",
+  "Mabel",
+  "Pearl",
+  "Marjorie",
+  "Gertrude",
+  "Lucille",
+  "Verity",
+  "Cordelia",
+  "Imelda",
+  "Reinhold",
+  "Bjorn",
+  "Sven",
+  "Ulrich",
+  "Klaus",
+  "Hans",
+  "Dieter",
+] as const;
