+++ src/shared/band-name-generator/word-pools/band-words.ts
@@ -1,0 +1,374 @@
+/** Adjectives, descriptors, verbs, and name suffixes for the band-name generator. */
+
+/** Weird, evocative adjectives suitable for metal/rock band names */
+export const BAND_ADJECTIVES = [
+  "Brutal",
+  "Equine",
+  "Mystic",
+  "Screaming",
+  "Savage",
+  "Eternal",
+  "Forbidden",
+  "Cursed",
+  "Crystal",
+  "Velvet",
+  "Iron",
+  "Black",
+  "Crimson",
+  "Ghastly",
+  "Howling",
+  "Roaring",
+  "Whispering",
+  "Burning",
+  "Frozen",
+  "Molten",
+  "Electric",
+  "Toxic",
+  "Radioactive",
+  "Phantom",
+  "Spectral",
+  "Skeletal",
+  "Demonic",
+  "Angelic",
+  "Holy",
+  "Unholy",
+  "Profane",
+  "Sacred",
+  "Vengeful",
+  "Wretched",
+  "Tragic",
+  "Tortured",
+  "Twisted",
+  "Bleeding",
+  "Pale",
+  "Withered",
+  "Reanimated",
+  "Drunken",
+  "Feral",
+  "Rabid",
+  "Mighty",
+  "Tyrannical",
+  "Imperial",
+  "Royal",
+  "Divine",
+  "Apocalyptic",
+  "Cosmic",
+  "Lunar",
+  "Solar",
+  "Stellar",
+  "Astral",
+  "Voidborn",
+  "Infernal",
+  "Diabolical",
+  "Sinister",
+  "Cataclysmic",
+  "Volcanic",
+  "Glacial",
+  "Granite",
+  "Marbled",
+  "Gilded",
+  "Tarnished",
+  "Mercurial",
+  "Mongolian",
+  "Jurassic",
+  "Chrome",
+  "Diesel",
+  "Atomic",
+  "Neon",
+  "Plastic",
+  "Wooden",
+  "Rusted",
+  "Festering",
+  "Slithering",
+  "Glowing",
+  "Magnetic",
+  "Thunderous",
+  "Subterranean",
+  "Hyperborean",
+  "Perpetual",
+  "Hooded",
+  "Velour",
+  "Magic",
+  "Naked",
+  "Hidden",
+  "Drowning",
+  "Choking",
+  "Greasy",
+  "Dusty",
+  "Rotting",
+  "Yelping",
+  "Polite",
+  "Ancient",
+  "Vintage",
+  "Final",
+  "Last",
+  "Crawling",
+  "Lurking",
+  "Hovering",
+  "Disgraced",
+  "Disowned",
+  "Defeated",
+  "Vanquished",
+  "Resurrected",
+  "Banished",
+  "Exiled",
+  "Wandering",
+  "Lost",
+  "Hunted",
+  "Haunted",
+  "Possessed",
+  "Anointed",
+  "Forsworn",
+  "Forgotten",
+  "Buried",
+  "Entombed",
+  "Enchanted",
+  "Bewitched",
+  "Spellbound",
+  "Hexed",
+  "Doomed",
+  "Damned",
+  "Glorious",
+  "Sublime",
+  "Foul",
+  "Putrid",
+  "Rancid",
+  "Acrid",
+  "Smouldering",
+  "Sizzling",
+  "Roasting",
+  "Boiling",
+  "Stewing",
+  "Bubbling",
+  "Sputtering",
+  "Crackling",
+  "Flickering",
+  "Faltering",
+  "Trembling",
+  "Shuddering",
+  "Quivering",
+  "Quaking",
+  "Wailing",
+  "Snarling",
+  "Gnashing",
+  "Spitting",
+  "Hissing",
+  "Buzzing",
+  "Droning",
+  "Squawking",
+  "Squelching",
+  "Gurgling",
+  "Slurping",
+  "Sticky",
+  "Knotted",
+  "Tangled",
+  "Crooked",
+  "Crumpled",
+  "Crumbling",
+  "Shattered",
+  "Splintered",
+  "Cracked",
+  "Fractured",
+  "Ruined",
+  "Decrepit",
+  "Derelict",
+  "Forlorn",
+  "Bereft",
+  "Abandoned",
+  "Forsaken",
+  "Regrettable",
+  "Ill-Advised",
+  "Untrustworthy",
+  "Suspicious",
+  "Unhinged",
+  "Bewildered",
+  "Befuddled",
+  "Flummoxed",
+  "Ungovernable",
+  "Inconsolable",
+  "Unbothered",
+  "Begrudging",
+  "Intolerable",
+  "Embarrassing",
+  "Implausible",
+  "Improbable",
+  "Unspeakable",
+  "Inexplicable",
+  "Unrepentant",
+  "Belligerent",
+  "Insatiable",
+  "Inevitable",
+  "Indomitable",
+  "Insufferable",
+  "Impervious",
+  "Imperishable",
+  "Incandescent",
+  "Indignant",
+  "Truculent",
+  "Petulant",
+  "Diminutive",
+  "Gargantuan",
+  "Colossal",
+  "Lopsided",
+  "Bedraggled",
+  "Bewhiskered",
+  "Bespectacled",
+  "Cantankerous",
+  "Pernicious",
+  "Lugubrious",
+  "Saccharine",
+  "Melodious",
+  "Pugilistic",
+  "Sclerotic",
+] as const;
+
+/** Collective nouns for "the band's reputation" */
+export const BAND_DESCRIPTORS = [
+  "pioneers",
+  "warriors",
+  "legends",
+  "veterans",
+  "prophets",
+  "godfathers",
+  "stalwarts",
+  "diehards",
+  "originators",
+  "cult heroes",
+  "mavericks",
+  "renegades",
+  "outlaws",
+  "old hands",
+  "elder statesmen",
+  "scene leaders",
+  "underdogs",
+  "scene-stealers",
+  "deep-cut specialists",
+  "self-styled visionaries",
+  "improbable survivors",
+  "eternal underdogs",
+  "long-suffering road dogs",
+  "near-mythical figures",
+  "barely-functional collective",
+  "loosely-affiliated crew",
+  "occasionally-touring side project",
+  "perennially-overlooked outfit",
+  "criminally-undercelebrated trio",
+  "begrudgingly-respected veterans",
+] as const;
+
+/** Verbs the band itself does — used in "{bandDescriptor} {bandVerb} ..." patterns */
+export const BAND_VERBS = [
+  "reunite",
+  "return",
+  "regroup",
+  "reform",
+  "ascend",
+  "descend",
+  "emerge",
+  "materialise",
+  "awaken",
+  "surface",
+  "converge",
+  "rally",
+  "assemble",
+  "manifest",
+  "reappear",
+  "convene",
+  "reconvene",
+  "stagger back",
+  "shamble forth",
+  "limp onstage",
+  "march in",
+  "stomp out",
+  "drift in",
+  "crash through the door",
+] as const;
+
+/** Tail phrases that turn a noun phrase into a fully-realised band name */
+export const BAND_SUFFIXES = [
+  "of Doom",
+  "of Mercy",
+  "of Sorrow",
+  "of the Damned",
+  "of the Forsaken",
+  "and Friends",
+  "Reunited",
+  "Unleashed",
+  "Resurrected",
+  "Returns",
+  "Forever",
+  "Reborn",
+  "Triumphant",
+  "Ascendant",
+  "of the Apocalypse",
+  "of the Void",
+  "of the Storm",
+  "of the Abyss",
+  "of the North",
+  "of Chaos",
+  "of Madness",
+  "of Eternity",
+  "of the Crypt",
+  "and the Disciples",
+  "and the Acolytes",
+  "and the Pilgrims",
+  "and Sons",
+  "and Daughters",
+  "Experience",
+  "Project",
+  "Collective",
+  "Brigade",
+  "Brotherhood",
+  "Sisterhood",
+  "Cult",
+  "Coven",
+  "Convocation",
+  "Conspiracy",
+  "Manifesto",
+  "Revival",
+  "Live in Concert",
+  "MMXXVI",
+  "of the South",
+  "of the East",
+  "of the West",
+  "of Yesterday",
+  "of Tomorrow",
+  "of the Underworld",
+  "of the Overworld",
+  "of the Marshlands",
+  "of the Highlands",
+  "of the Lowlands",
+  "of the Swamp",
+  "of the Steppe",
+  "of the Tundra",
+  "of the Ravine",
+  "Reanimated",
+  "Refurbished",
+  "Recommissioned",
+  "Reconvened",
+  "Re-Released",
+  "and the Quartermasters",
+  "and the Apothecaries",
+  "and the Notaries",
+  "and the Clerks",
+  "and the Custodians",
+  "Quintet",
+  "Sextet",
+  "Septet",
+  "Orchestra",
+  "Ensemble",
+  "Society",
+  "Guild",
+  "Syndicate",
+  "Authority",
+  "Committee",
+  "Tribunal",
+  "Inquest",
+  "Convention",
+  "Roadshow",
+  "Caravan",
+  "Pageant",
+  "Processional",
+  "Wake",
+  "Vigil",
+] as const;
