+++ src/shared/band-name-generator/word-pools/music.ts
@@ -1,0 +1,287 @@
+/** Genres, festivals, listing types, and tour flavour for gig descriptions. */
+
+/** Sub-genres used inside gig-description templates */
+export const GENRES = [
+  "doom",
+  "sludge",
+  "prog",
+  "hardcore",
+  "glam",
+  "thrash",
+  "death-metal",
+  "black-metal",
+  "stoner-rock",
+  "folk-metal",
+  "industrial",
+  "symphonic",
+  "hair-metal",
+  "gothic",
+  "drone",
+  "speed-metal",
+  "power-metal",
+  "psychobilly",
+  "darkwave",
+  "post-punk",
+  "shoegaze",
+  "grindcore",
+  "blackened-jazz",
+  "viking-metal",
+  "pirate-metal",
+  "doom-jazz",
+  "djent",
+  "crust-punk",
+  "dream-pop",
+  "math-rock",
+  "noise-rock",
+  "krautrock",
+  "post-rock",
+  "post-metal",
+  "blackgaze",
+  "deathcore",
+  "metalcore",
+  "groovecore",
+  "nu-metal",
+  "rap-metal",
+  "country-metal",
+  "celtic-metal",
+  "medieval-metal",
+  "baroque-metal",
+  "circus-metal",
+  "chiptune-metal",
+  "synthwave",
+  "electroclash",
+  "horror-punk",
+  "cowpunk",
+  "skiffle-metal",
+  "dungeon-synth",
+  "bedroom-doom",
+  "garage-thrash",
+  "polka-grind",
+  "screamo",
+  "emo-violence",
+  "math-core",
+  "post-hardcore",
+  "djent-folk",
+  "chamber-doom",
+  "zydeco-thrash",
+  "sea-shanty-core",
+] as const;
+
+/** Themes for crossover/novelty acts */
+export const CROSSOVERS = [
+  "sea-shanty",
+  "druid",
+  "viking",
+  "pirate",
+  "robot",
+  "ghost",
+  "morris-dance",
+  "bagpipe",
+  "yodelling",
+  "yacht-rock",
+  "monster-truck",
+  "sci-fi",
+  "wild-west",
+  "Tudor-court",
+  "Edwardian-music-hall",
+  "Roman-legion",
+  "samurai",
+  "lumberjack",
+  "tea-party",
+  "coal-miner",
+  "deep-sea-trawler",
+  "haunted-circus",
+  "post-apocalyptic",
+  "Cold-War",
+  "Steampunk",
+  "Cyberpunk",
+  "renaissance-fair",
+  "1970s-prog",
+  "1980s-disco",
+  "1990s-rave",
+  "early-noughties-emo",
+  "small-claims-court",
+  "town-council",
+  "garden-centre",
+] as const;
+
+/** Outdoor / listing-style venue archetypes */
+export const FESTIVAL_TYPES = [
+  "Festival",
+  "Fest",
+  "Fair",
+  "Carnival",
+  "Convention",
+  "Summit",
+  "Showcase",
+  "Massacre",
+  "Slaughter",
+  "Bash",
+  "Jamboree",
+  "Riot",
+  "Reckoning",
+  "Pilgrimage",
+  "Procession",
+  "Reunion",
+  "Tour",
+  "Crusade",
+  "Onslaught",
+  "Weekender",
+  "All-Dayer",
+  "Knees-Up",
+  "Hoedown",
+  "Stomp",
+  "Throwdown",
+  "Free-For-All",
+  "Marathon",
+  "Wakefield",
+  "Vigil",
+  "Spectacle",
+  "Extravaganza",
+  "Gala",
+  "Soirée",
+  "Roadshow",
+  "Symposium",
+  "Conclave",
+  "Assembly",
+  "Gathering",
+  "Reunion Tour",
+  "Exhibition",
+  "Tournament",
+  "Olympiad",
+] as const;
+
+/** What the listing itself is — sits where words like "show" / "matinee" used to be hard-coded */
+export const LISTING_TYPES = [
+  "night",
+  "matinee",
+  "show",
+  "evening",
+  "gig",
+  "set",
+  "recital",
+  "séance",
+  "summoning",
+  "ritual",
+  "vigil",
+  "broadcast",
+  "takeover",
+  "residency",
+  "occupation",
+  "livestream",
+  "riot",
+  "jamboree",
+  "knees-up",
+  "happening",
+  "spectacle",
+  "exhibition",
+  "convocation",
+  "wake",
+  "send-off",
+  "reckoning",
+  "exorcism",
+] as const;
+
+/** Verbs for "the band ___ed the venue" */
+export const NOISE_VERBS = [
+  "invade",
+  "descend on",
+  "tear through",
+  "assault",
+  "demolish",
+  "conquer",
+  "devastate",
+  "obliterate",
+  "occupy",
+  "set fire to",
+  "lay siege to",
+  "blockade",
+  "annex",
+  "decommission",
+  "rearrange",
+  "renegotiate",
+  "remodel",
+  "reupholster",
+  "redecorate",
+  "fumigate",
+  "rezone",
+  "dismantle",
+  "reassemble",
+  "abscond from",
+  "infiltrate",
+  "depopulate",
+  "haunt",
+  "stalk",
+  "loiter in",
+  "graffiti",
+  "monopolise",
+  "outlaw",
+] as const;
+
+/** Hyperbolic adjective phrases for sensory descriptions */
+export const INTENSITIES = [
+  "ear-splitting",
+  "brain-melting",
+  "head-splitting",
+  "ground-shaking",
+  "soul-crushing",
+  "deafening",
+  "earth-shattering",
+  "skin-peeling",
+  "bone-rattling",
+  "chest-thumping",
+  "thoroughly inadvisable",
+  "barely-legal",
+  "trouser-flapping",
+  "hair-raising",
+  "spleen-loosening",
+  "neighbour-bothering",
+  "windscreen-shattering",
+  "pavement-cracking",
+  "ceiling-loosening",
+  "double-glazing-defeating",
+  "rib-rattling",
+  "cardigan-blowing",
+  "biscuit-tin-disturbing",
+  "regrettably enthusiastic",
+  "unreasonably cheerful",
+  "borderline-dangerous",
+  "questionably-permitted",
+  "absurdly-loud",
+  "unaccountably-aggressive",
+  "knee-buckling",
+  "teeth-shaking",
+  "moustache-curling",
+  "ozone-depleting",
+  "hedge-trimming",
+  "impressively-illegal",
+] as const;
+
+/** Adjectives that qualify a "tour" or one-off comeback show */
+export const TOUR_ADJECTIVES = [
+  "farewell",
+  "reunion",
+  "comeback",
+  "retrospective",
+  "anniversary",
+  "definitely-the-last",
+  "victory",
+  "redemption",
+  "vengeance",
+  "valedictory",
+  "absolutely-final",
+  "really-really-last",
+  "we-mean-it-this-time",
+  "first and last",
+  "swansong",
+  "homecoming",
+  "tearful",
+  "long-awaited",
+  "long-promised",
+  "long-overdue",
+  "mostly-original-lineup",
+  "approximately-original-lineup",
+  "one-night-only",
+  "two-nights-only",
+  "open-ended",
+] as const;
