+++ src/shared/band-name-generator/word-pools/shows.ts
@@ -1,0 +1,300 @@
+/** Show items, prizes, rules, and the small filler pools (numbers, times, connectors). */
+
+/** Concrete things you'd see / smell / regret at a gig */
+export const SHOW_ITEMS = [
+  "denim",
+  "leather",
+  "mead",
+  "regret",
+  "glory",
+  "smoke",
+  "lasers",
+  "dry ice",
+  "bandanas",
+  "studs",
+  "shredding solos",
+  "blast beats",
+  "double kicks",
+  "fog machines",
+  "questionable lyrics",
+  "obscure cover versions",
+  "extended drum solos",
+  "merch you'll regret on Monday",
+  "improvised pyrotechnics",
+  "spilled beer",
+  "spandex",
+  "corpse paint",
+  "bullet belts",
+  "leather kilts",
+  "body glitter",
+  "patchouli",
+  "warm lager",
+  "cold sausage rolls",
+  "merchandise queues",
+  "guitar picks",
+  "drumstick souvenirs",
+  "a single rogue bagpipe",
+  "exhausted sound engineers",
+  "questionable hairlines",
+  "ill-advised tattoos",
+  "second-hand smoke",
+  "third-hand stories",
+  "sweat-stained setlists",
+  "audible feedback",
+  "inaudible vocals",
+  "broken cymbals",
+  "duct tape",
+  "split jeans",
+  "torn t-shirts",
+  "missing teeth",
+  "discounted earplugs",
+  "questionable life choices",
+  "everyone's last shred of dignity",
+  "free posters",
+  "expired pizza vouchers",
+  "improvised mosh choreography",
+] as const;
+
+/** Karaoke / dress-up prize categories */
+export const PRIZES = [
+  "biggest hair",
+  "loudest scream",
+  "best devil horns",
+  "most convincing growl",
+  "longest air-guitar solo",
+  "deepest voice",
+  "most denim worn at once",
+  "most authentic mullet",
+  "most enthusiastic windmill arm",
+  "best fake corpse-paint",
+  "loudest unaccompanied howl",
+  "longest sustained note",
+  "most accurate Lemmy impression",
+  "fastest tambourine work",
+  "most dramatic hair flip",
+  "most-improved scream since last time",
+  "best-disguised vocal warm-up",
+  "most committed knee-slide",
+  "most emphatic horns-up gesture",
+  "best Joe-Cocker-style face",
+] as const;
+
+/** Things that are absolutely not allowed at this gig */
+export const PROHIBITIONS = [
+  "stage diving onto the pizza",
+  "phones above shoulder height",
+  "vuvuzelas",
+  "glitter cannons",
+  "throwing pints",
+  "uninvited cowbell",
+  "amateur pyrotechnics",
+  "crowdsurfing inflatables",
+  "audience requests",
+  "pretending to know the lyrics",
+  "shouting Wonderwall",
+  "encore-baiting",
+  "bringing your own setlist",
+  "open-toed footwear in the pit",
+  "drum solos by attendees",
+  "tribute-act T-shirts",
+  "loud crisp packets",
+  "moshing during ballads",
+  "talking during ballads",
+  "filming with two phones at once",
+  "selfie sticks of any length",
+  "questioning the headliner's setlist",
+  "improvised lyrics",
+  "queue-jumping at merch",
+  "uninvited stage invasions",
+  "balloon animals",
+  "vapes the size of a kettle",
+  "outdoor flares",
+  "professional cameras without a pass",
+  "harmonicas brought from home",
+  "negotiating refunds mid-set",
+  "audible sighs during solos",
+  "inappropriately enthusiastic kazoo work",
+] as const;
+
+/** Items of clothing or attire that are notionally compulsory */
+export const REQUIREMENTS = [
+  "sequins",
+  "leather",
+  "denim",
+  "mascara",
+  "hairspray",
+  "shoulder pads",
+  "bullet belts",
+  "tasselled jackets",
+  "shredded jeans",
+  "spandex",
+  "corpse paint",
+  "studded gauntlets",
+  "pirate hats",
+  "viking helmets",
+  "novelty cravats",
+  "ill-fitting capes",
+  "platform boots",
+  "bandanas (any colour)",
+  "ironic moustaches",
+  "elaborate eyeliner",
+  "battle vests",
+  "back patches",
+  "fingerless gloves",
+  "wraparound sunglasses",
+  "kohl-rimmed eyes",
+  "dramatic neckwear",
+  "regrettable haircuts",
+  "tour T-shirts (preferably someone else's)",
+] as const;
+
+/** Predictable consequences of attending the gig */
+export const AUDIENCE_OUTCOMES = [
+  "regret",
+  "glory",
+  "tinnitus",
+  "sweat",
+  "tears",
+  "mild concern",
+  "confusion",
+  "broken eardrums",
+  "lost dignity",
+  "hoarse voices",
+  "lost car keys",
+  "aching necks",
+  "mismatched footwear",
+  "sticky floors",
+  "questionable bruises",
+  "slightly damp denim",
+  "sudden tinnitus",
+  "minor brain fog",
+  "unexplained leaves in pockets",
+  "inappropriate enthusiasm",
+  "newfound dietary requirements",
+  "missed last buses",
+  "incoherent stories",
+  "second-hand glitter",
+] as const;
+
+/** Audience-eligibility blurbs */
+export const AGE_NOTES = [
+  "all ages welcome",
+  "18+ only",
+  "14+ with parent",
+  "family-friendly",
+  "parental advisory in effect",
+  "strictly grown-ups",
+  "over-30s strongly encouraged",
+  "no under-fives without ear defenders",
+  "pets on leads only",
+  "ID checks at the door",
+  "members and accompanied guests",
+  "carer admitted free",
+  "no babes in arms after 9pm",
+  "well-behaved teens permitted",
+  "responsible adults preferred",
+] as const;
+
+/** Prepositional phrases connecting clauses without being entirely hard-coded */
+export const CONNECTORS = [
+  "with",
+  "featuring",
+  "accompanied by",
+  "fuelled by",
+  "propelled by",
+  "soundtracked by",
+  "garnished with",
+  "interrupted by",
+  "haunted by",
+  "supported by",
+  "underwritten by",
+  "invigilated by",
+  "officiated by",
+  "narrated by",
+  "punctuated by",
+  "occasionally upstaged by",
+  "loosely supervised by",
+  "barely tolerated by",
+] as const;
+
+/** Spelled-out small numbers for descriptions */
+export const NUMBER_WORDS = [
+  "two",
+  "three",
+  "four",
+  "five",
+  "six",
+  "seven",
+  "eight",
+  "nine",
+  "ten",
+  "eleven",
+  "twelve",
+  "thirteen",
+  "slistingeen",
+  "twenty-one",
+  "thirty",
+  "an alarming number of",
+  "an undisclosed number of",
+  "no fewer than four",
+  "approximately seven",
+  "exactly one",
+] as const;
+
+/** Wildly out-of-place instruments for novelty descriptions */
+export const ODD_INSTRUMENTS = [
+  "theremin",
+  "ukulele",
+  "bagpipes",
+  "hurdy-gurdy",
+  "kazoo",
+  "glockenspiel",
+  "accordion",
+  "harp",
+  "sitar",
+  "mandolin",
+  "stylophone",
+  "didgeridoo",
+  "vacuum cleaner",
+  "Casio keyboard",
+  "harpsichord",
+  "calliope",
+  "musical saw",
+  "spoons",
+  "washboard",
+  "jaw harp",
+  "ocarina",
+  "pan flute",
+  "alphorn",
+  "glass harmonica",
+  "lap steel",
+  "slide whistle",
+  "swanee whistle",
+  "wobble board",
+  "Stylophone Beatbox",
+  "single bagpipe drone",
+  "broken Wurlitzer",
+  "ironing board with strings",
+  "rubber-band orchestra",
+  "tuned biscuit tin",
+  "borrowed cello",
+  "untuned banjo",
+  "kettle and spoon",
+  "vintage ondes Martenot",
+  "bicycle bell choir",
+  "amplified beard",
+] as const;
+
+/** Common gig start times (24h-style hours) */
+export const TIMES = [
+  "6",
+  "7",
+  "7:30",
+  "8",
+  "8:30",
+  "9",
+  "9:30",
+  "10",
+  "11",
+  "midnight",
+] as const;
