+++ src/shared/band-name-generator/word-pools/venues.ts
@@ -1,0 +1,136 @@
+/** Venue names, weird venues, and building states. */
+
+/** Indoor venue archetypes — concrete buildings/rooms */
+export const VENUE_TYPES = [
+  "Arena",
+  "Academy",
+  "Stadium",
+  "Auditorium",
+  "Hall",
+  "Theatre",
+  "Coliseum",
+  "Amphitheatre",
+  "Pavilion",
+  "Forum",
+  "Dome",
+  "Centre",
+  "Roundhouse",
+  "Tabernacle",
+  "Sanctuary",
+  "Crypt",
+  "Vault",
+  "Bunker",
+  "Warehouse",
+  "Factory",
+  "Mill",
+  "Foundry",
+  "Quarry",
+  "Depot",
+  "Garage",
+  "Lounge",
+  "Club",
+  "Den",
+  "Pit",
+  "Chamber",
+  "Ballroom",
+  "Workingmen's Club",
+  "Social Club",
+  "Working Men's Institute",
+  "Bingo Hall",
+  "Snooker Hall",
+  "Boxing Gym",
+  "Scout Hut",
+  "Village Hall",
+  "Lecture Theatre",
+  "Cinema",
+  "Skating Rink",
+  "Roller Disco",
+  "Bowling Alley",
+  "Mausoleum",
+  "Crematorium",
+  "Conservatory",
+  "Pump House",
+  "Boiler Room",
+  "Engine Shed",
+  "Tram Shed",
+  "Toll House",
+  "Custom House",
+  "Drill Hall",
+  "Masonic Lodge",
+  "Reading Room",
+  "Tea Rooms",
+  "Assembly Rooms",
+  "Music Room",
+  "Practice Room",
+  "Rehearsal Studio",
+] as const;
+
+/** Improbable buildings to be "converted into" a venue */
+export const WEIRD_VENUES = [
+  "lift shaft",
+  "granary",
+  "bus depot",
+  "brewery vat",
+  "postal sorting office",
+  "cattle market",
+  "abandoned multi-storey",
+  "decommissioned chapel",
+  "1980s leisure centre",
+  "shipping container",
+  "village butcher's freezer",
+  "art-deco cinema",
+  "Victorian bathhouse",
+  "Edwardian bandstand",
+  "disused fire station",
+  "former post office",
+  "redundant church",
+  "out-of-service ferry",
+  "decommissioned lighthouse",
+  "abandoned underground station",
+  "WW2 pillbox",
+  "converted barn",
+  "haunted greenhouse",
+  "old reading room",
+  "empty Woolworths",
+  "former Blockbuster",
+  "former HMV",
+  "shut-down bowling alley",
+  "padlocked snooker hall",
+  "boarded-up pub",
+  "derelict library",
+  "scrapped trawler",
+  "retired narrowboat",
+  "carriage-works",
+  "pump house",
+  "abandoned railway tunnel",
+  "long-suffering scout hut",
+  "regrettable nightclub",
+  "decommissioned petrol station",
+] as const;
+
+/** Adjective phrases describing the state of a "weird venue" before being repurposed */
+export const BUILDING_STATES = [
+  "converted",
+  "repurposed",
+  "abandoned",
+  "decommissioned",
+  "gentrified",
+  "restored",
+  "haunted",
+  "condemned",
+  "derelict",
+  "half-finished",
+  "recently-evacuated",
+  "formerly-respectable",
+  "barely-legal",
+  "improbably-listed",
+  "informally-occupied",
+  "structurally-unsound",
+  "reluctantly-available",
+  "council-owned",
+  "privately-disowned",
+  "long-disused",
+  "freshly-tagged",
+  "improvisationally-rewired",
+  "inadvisably-reopened",
+] as const;
